feat: 인기 상품 목록 조회 API, Product 인기도 Long으로 변경

+++ src/main/java/com/process/clash/adapter/persistence/shop/product/ProductJpaEntity.java
@@ -1,7 +1,6 @@
 package com.process.clash.adapter.persistence.shop.product;
 
 import com.process.clash.adapter.persistence.shop.season.SeasonJpaEntity;
-import com.process.clash.domain.shop.product.entity.Product;
 import com.process.clash.domain.shop.product.enums.ProductCategory;
 import com.process.clash.domain.shop.product.enums.ProductGoodsType;
 import jakarta.persistence.*;
@@ -57,7 +56,7 @@ public class ProductJpaEntity {
     private String description;
 
     @Column(nullable = false)
-    private Integer popularity;
+    private Long popularity;
 
     @JoinColumn(name = "fk_season_id")
     @ManyToOne(fetch = FetchType.LAZY)
